[VW-377] chore: reorganized markdown helpers in new lib
