+++ src/features/inbox/agent/vex/context.ts
@@ -15,7 +15,9 @@ import {
   deviceGroupLabel,
   deviceGroupMatchingLabel,
   deviceGroupToMarkdown,
-} from "@/lib/string-utils";
+  type NoteRow,
+  renderNoteTarget,
+} from "@/lib/markdown";
 
 /** A baseline (device-group-matching-level) issue the agent may refine. */
 export type VexIssueContext = {
@@ -202,38 +204,6 @@ export async function gatherVexContext(
   });
 
   return { notificationId, markdown, issues };
-}
-
-type NoteRow = {
-  text: string;
-  status: string;
-  targetModel: string | null;
-  instanceId: string | null;
-};
-
-function renderNoteTarget(
-  note: NoteRow,
-  labels: {
-    assetLabel: Map<string, string>;
-    groupLabel: Map<string, string>;
-    matchingLabel: Map<string, string>;
-    cveById: Map<string, string>;
-  },
-): string {
-  if (note.status === "PERSISTENT") return "Persistent (hospital-wide)";
-  const id = note.instanceId ?? "";
-  switch (note.targetModel) {
-    case "ASSET":
-      return `Asset ${labels.assetLabel.get(id) ?? id} (id: ${id})`;
-    case "DEVICE_GROUP_MATCHING":
-      return `Matching ${labels.matchingLabel.get(id) ?? id}`;
-    case "VULNERABILITY":
-      return `Vulnerability ${labels.cveById.get(id) ?? id}`;
-    case "REMEDIATION":
-      return `Remediation ${id}`;
-    default:
-      return "Unknown target";
-  }
 }
 
 function renderVexPrompt(args: {
