+++ src/features/vulnerabilities/components/vulnerability-drawer.tsx
@@ -29,7 +29,7 @@ import {
 import type { UserRole } from "@/features/chat/utils";
 import { IssuesSidebarList } from "@/features/issues/components/issue";
 import { RemediationCard } from "@/features/remediations/components/remediations";
-import { deviceGroupMatchingLabel } from "@/lib/string-utils";
+import { deviceGroupMatchingLabel } from "@/lib/markdown";
 import type { VulnerabilityWithRelations } from "../types";
 
 // ============================================================================
