+++ src/lib/markdown/vulnerability.ts
@@ -1,0 +1,88 @@
+// Vulnerability → markdown renderer. Structurally typed so any Prisma payload
+// with these fields is compatible.
+
+import { deviceGroupMatchingsSummary } from "./device-group";
+import { type CanonicalRef, shortId, truncate } from "./shared";
+
+type DeviceGroupMatchingForMarkdown = {
+  vendor?: CanonicalRef;
+  product?: CanonicalRef;
+  version?: CanonicalRef;
+  versionRange?: string | null;
+};
+
+export interface VulnerabilityForMarkdown {
+  id: string;
+  cveId?: string | null;
+  severity: string;
+  priority?: string;
+  cvssScore?: number | null;
+  cvssVector?: string | null;
+  epss?: number | null;
+  inKEV?: boolean;
+  exploitUri?: string | null;
+  affectedComponents?: string[];
+  description?: string | null;
+  narrative?: string | null;
+  impact?: string | null;
+  deviceGroupMatchings?: DeviceGroupMatchingForMarkdown[];
+  remediations?: Array<{ id: string; description?: string | null }>;
+  issues?: Array<{
+    status: string;
+    asset: { id: string; hostname?: string | null; ip?: string | null } | null;
+  }>;
+}
+
+export function vulnerabilityToMarkdown(
+  v: VulnerabilityForMarkdown,
+  opts: { includeAssets?: boolean; includeRemediations?: boolean } = {
+    includeAssets: true,
+    includeRemediations: true,
+  },
+): string {
+  const lines = [
+    `### ${v.cveId ?? v.id} — ${v.severity}${v.priority ? `, priority: ${v.priority}` : ""}`,
+    `- **CVSS Score**: ${v.cvssScore ?? "N/A"}`,
+    `- **CVSS Vector**: ${v.cvssVector ?? "N/A"}`,
+    `- **EPSS**: ${v.epss != null ? `${(v.epss * 100).toFixed(2)}%` : "N/A"}`,
+    `- **In CISA KEV**: ${v.inKEV ? "Yes" : "No"}`,
+  ];
+
+  if (v.affectedComponents && v.affectedComponents.length > 0) {
+    lines.push(`- **Affected Components**: ${v.affectedComponents.join(", ")}`);
+  }
+
+  if (v.deviceGroupMatchings && v.deviceGroupMatchings.length > 0) {
+    lines.push(
+      `- **Affected Products**: ${deviceGroupMatchingsSummary(v.deviceGroupMatchings)}`,
+    );
+  }
+
+  if (v.description) lines.push(`- **Description**: ${v.description}`);
+  if (v.narrative) lines.push(`- **Exploit Narrative**: ${v.narrative}`);
+  if (v.impact) lines.push(`- **Clinical Impact**: ${v.impact}`);
+
+  if (opts.includeAssets && v.issues && v.issues.length > 0) {
+    lines.push(`- **Affected Assets** (${v.issues.length}):`);
+    for (const issue of v.issues) {
+      if (!issue.asset) {
+        lines.push(`  - device group — issue status: ${issue.status}`);
+        continue;
+      }
+      const label = issue.asset.hostname ?? issue.asset.ip ?? issue.asset.id;
+      lines.push(
+        `  - ${label} (${shortId(issue.asset.id)}) — issue status: ${issue.status}`,
+      );
+    }
+  }
+
+  if (opts.includeRemediations && v.remediations && v.remediations.length > 0) {
+    lines.push(`- **Available Remediations** (${v.remediations.length}):`);
+    for (const r of v.remediations) {
+      const desc = r.description ? ` — ${truncate(r.description, 80)}` : "";
+      lines.push(`  - rem-${shortId(r.id)}${desc}`);
+    }
+  }
+
+  return lines.join("\n");
+}
